Move export view functions to their own file
