+++ core/views/__init__.py
@@ -11,3 +11,4 @@ from .published import *
 from .record_identifer_transformation_scenario import *
 from .stateio import *
 from .background_tasks import *
+from .export import *
